refactor(ricer::cli)!: rename `ui` module to `cli`
I really do not intend on implementing anything other than a CLI for Ricer.
Thus, make this module match my intentions a little more.
